refactor: split config modules and stabilize consistency checks
