+++ src/config/index.js
@@ -1,0 +1,3 @@
+export * from './params.js';
+export * from './dev-ui.js';
+export * from './dev-registry.js';
