+++ tests/config-consistency.test.js
@@ -1,16 +1,17 @@
 /**
  * config-consistency.test.js
- * 
+ *
  * kesson-space 設定値の整合性テスト
  * ブラウザ不要 — Node.js で実行可能
- * 
+ *
  * 実行: node tests/config-consistency.test.js
- * 
+ *
  * 検証項目:
- * 1. dev-panel の default 値が config オブジェクトと一致
- * 2. シェーダー uniform 初期値が config と一致
- * 3. i18n の ja/en キー構造一致
- * 4. 未使用ファイルが削除済み
+ * 1. config 分割モジュールの存在と主要オブジェクト定義
+ * 2. dev-ui default 値が config パラメータ参照を維持していること
+ * 3. dev-registry の構造整合（toggle重複なし含む）
+ * 4. i18n の ja/en キー構造一致
+ * 5. 不要ファイルが残っていないこと
  */
 
 import { readFileSync, existsSync } from 'fs';
@@ -18,9 +19,10 @@ import { resolve, dirname } from 'path';
 import { fileURLToPath } from 'url';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
-const SRC = resolve(__dirname, '..', 'src');
+const ROOT = resolve(__dirname, '..');
+const SRC = resolve(ROOT, 'src');
+const CONFIG_DIR = resolve(SRC, 'config');
 
-// --- ユーティリティ ---
 let passed = 0;
 let failed = 0;
 const failures = [];
@@ -31,140 +33,130 @@ function assert(condition, message) {
     } else {
         failed++;
         failures.push(message);
-        console.error(`  \u2717 ${message}`);
+        console.error(`  ✗ ${message}`);
     }
 }
 
 function section(name) {
-    console.log(`\n\u2500\u2500 ${name} \u2500\u2500`);
+    console.log(`\n── ${name} ──`);
 }
 
-function readSrc(relPath) {
-    return readFileSync(resolve(SRC, relPath), 'utf-8');
+function read(path) {
+    return readFileSync(path, 'utf-8');
 }
-
-// ========================================
-// Test 1: config.js のパラメータ抽出
-// ========================================
-section('1. config.js \u30d1\u30e9\u30e1\u30fc\u30bf\u89e3\u6790');
-
-const configSrc = readSrc('config.js');
 
 function extractObject(src, name) {
     const regex = new RegExp(`export\\s+const\\s+${name}\\s*=\\s*\\{([^}]+)\\}`, 's');
     const match = src.match(regex);
     if (!match) return null;
+
     const body = match[1];
     const result = {};
     const lines = body.split('\n');
+
     for (const line of lines) {
-        const m = line.match(/^\s*(\w+)\s*:\s*([\d.+-]+)/);
-        if (m) result[m[1]] = parseFloat(m[2]);
-        const bm = line.match(/^\s*(\w+)\s*:\s*(true|false)/);
-        if (bm) result[bm[1]] = bm[2] === 'true';
+        const num = line.match(/^\s*(\w+)\s*:\s*([\d.+-]+)/);
+        if (num) result[num[1]] = Number.parseFloat(num[2]);
+
+        const bool = line.match(/^\s*(\w+)\s*:\s*(true|false)/);
+        if (bool) result[bool[1]] = bool[2] === 'true';
     }
+
     return result;
 }
 
-const sceneParams = extractObject(configSrc, 'sceneParams');
-const fluidParams = extractObject(configSrc, 'fluidParams');
-const distortionParams = extractObject(configSrc, 'distortionParams');
-const toggles = extractObject(configSrc, 'toggles');
-const breathConfig = extractObject(configSrc, 'breathConfig');
+section('1. config 分割モジュール確認');
 
-assert(sceneParams !== null, 'sceneParams \u304c config.js \u306b\u5b58\u5728\u3059\u308b');
-assert(fluidParams !== null, 'fluidParams \u304c config.js \u306b\u5b58\u5728\u3059\u308b');
-assert(distortionParams !== null, 'distortionParams \u304c config.js \u306b\u5b58\u5728\u3059\u308b');
-assert(toggles !== null, 'toggles \u304c config.js \u306b\u5b58\u5728\u3059\u308b');
-assert(breathConfig !== null, 'breathConfig \u304c config.js \u306b\u5b58\u5728\u3059\u308b');
+const configEntryPath = resolve(SRC, 'config.js');
+const configIndexPath = resolve(CONFIG_DIR, 'index.js');
+const paramsPath = resolve(CONFIG_DIR, 'params.js');
+const devUiPath = resolve(CONFIG_DIR, 'dev-ui.js');
+const devRegistryPath = resolve(CONFIG_DIR, 'dev-registry.js');
+
+assert(existsSync(configEntryPath), 'src/config.js が存在する');
+assert(existsSync(configIndexPath), 'src/config/index.js が存在する');
+assert(existsSync(paramsPath), 'src/config/params.js が存在する');
+assert(existsSync(devUiPath), 'src/config/dev-ui.js が存在する');
+assert(existsSync(devRegistryPath), 'src/config/dev-registry.js が存在する');
+
+const configEntrySrc = read(configEntryPath);
+const paramsSrc = read(paramsPath);
+const devUiSrc = read(devUiPath);
+const devRegistrySrc = read(devRegistryPath);
+
+assert(/export\s+\*\s+from\s+'\.\/config\/index\.js'/.test(configEntrySrc),
+    'src/config.js が config/index.js を再エクスポートしている');
+
+const sceneParams = extractObject(paramsSrc, 'sceneParams');
+const fluidParams = extractObject(paramsSrc, 'fluidParams');
+const distortionParams = extractObject(paramsSrc, 'distortionParams');
+const toggles = extractObject(paramsSrc, 'toggles');
+const breathConfig = extractObject(paramsSrc, 'breathConfig');
+
+assert(sceneParams !== null, 'sceneParams が params.js に存在する');
+assert(fluidParams !== null, 'fluidParams が params.js に存在する');
+assert(distortionParams !== null, 'distortionParams が params.js に存在する');
+assert(toggles !== null, 'toggles が params.js に存在する');
+assert(breathConfig !== null, 'breathConfig が params.js に存在する');
 
 console.log(`  sceneParams: ${Object.keys(sceneParams || {}).length} keys`);
 console.log(`  fluidParams: ${Object.keys(fluidParams || {}).length} keys`);
 console.log(`  distortionParams: ${Object.keys(distortionParams || {}).length} keys`);
 
-// ========================================
-// Test 2: dev-panel のdefaultがconfigを参照しているか
-// ========================================
-section('2. dev-panel default \u53c2\u7167\u30c1\u30a7\u30c3\u30af');
-
-const devPanelSrc = readSrc('dev-panel.js');
+section('2. dev-ui default 参照チェック');
 
 const configRefs = [
-    { pattern: /default:\s*fluidParams\.force/,    name: 'fluidForce \u2192 fluidParams.force' },
-    { pattern: /default:\s*fluidParams\.curl/,     name: 'fluidCurl \u2192 fluidParams.curl' },
-    { pattern: /default:\s*fluidParams\.decay/,    name: 'fluidDecay \u2192 fluidParams.decay' },
-    { pattern: /default:\s*fluidParams\.radius/,   name: 'fluidRadius \u2192 fluidParams.radius' },
-    { pattern: /default:\s*fluidParams\.influence/, name: 'fluidInfluence \u2192 fluidParams.influence' },
-    { pattern: /default:\s*distortionParams\.strength/,  name: 'distStrength \u2192 distortionParams.strength' },
-    { pattern: /default:\s*distortionParams\.aberration/, name: 'distAberration \u2192 distortionParams.aberration' },
-    { pattern: /default:\s*distortionParams\.haloColorR/, name: 'haloColorR \u2192 distortionParams.haloColorR' },
-    { pattern: /default:\s*sceneParams\.brightness/, name: 'brightness \u2192 sceneParams.brightness' },
-    { pattern: /default:\s*sceneParams\.tintR/,     name: 'tintR \u2192 sceneParams.tintR' },
-    { pattern: /default:\s*breathConfig\.period/,   name: 'period \u2192 breathConfig.period' },
+    { pattern: /default:\s*fluidParams\.force/, name: 'fluidForce → fluidParams.force' },
+    { pattern: /default:\s*fluidParams\.curl/, name: 'fluidCurl → fluidParams.curl' },
+    { pattern: /default:\s*fluidParams\.decay/, name: 'fluidDecay → fluidParams.decay' },
+    { pattern: /default:\s*fluidParams\.radius/, name: 'fluidRadius → fluidParams.radius' },
+    { pattern: /default:\s*fluidParams\.influence/, name: 'fluidInfluence → fluidParams.influence' },
+    { pattern: /default:\s*distortionParams\.strength/, name: 'distStrength → distortionParams.strength' },
+    { pattern: /default:\s*distortionParams\.aberration/, name: 'distAberration → distortionParams.aberration' },
+    { pattern: /default:\s*distortionParams\.haloColorR/, name: 'haloColorR → distortionParams.haloColorR' },
+    { pattern: /default:\s*sceneParams\.brightness/, name: 'brightness → sceneParams.brightness' },
+    { pattern: /default:\s*sceneParams\.tintR/, name: 'tintR → sceneParams.tintR' },
+    { pattern: /default:\s*breathConfig\.period/, name: 'period → breathConfig.period' },
 ];
 
 for (const ref of configRefs) {
-    assert(ref.pattern.test(devPanelSrc), `dev-panel: ${ref.name}`);
+    assert(ref.pattern.test(devUiSrc), `dev-ui: ${ref.name}`);
 }
 
-// ========================================
-// Test 3: シェーダー uniform 参照チェック
-// ========================================
-section('3. \u30b7\u30a7\u30fc\u30c0\u30fc uniform \u53c2\u7167\u30c1\u30a7\u30c3\u30af');
+section('3. dev-registry / toggle 構造チェック');
 
-const fluidFieldSrc = readSrc('shaders/fluid-field.js');
-const distortionSrc = readSrc('shaders/distortion-pass.js');
-const kessonSrc = readSrc('shaders/kesson.js');
+assert(/export\s+const\s+DEV_TOGGLES\s*=\s*\[/.test(devUiSrc), 'DEV_TOGGLES が定義されている');
+assert(/export\s+const\s+DEV_SECTIONS\s*=\s*\[/.test(devUiSrc), 'DEV_SECTIONS が定義されている');
+assert(/export\s+const\s+DEV_PARAM_REGISTRY/.test(devRegistrySrc), 'DEV_PARAM_REGISTRY が定義されている');
+assert(/registry\[toggle\.key\]/.test(devRegistrySrc), 'toggle エントリが registry に反映される');
+assert(/DEV_SECTIONS\.forEach/.test(devRegistrySrc), 'DEV_SECTIONS を元に registry を構築している');
 
-// includes() ベース — 正規表現リテラル内の '/' エスケープ問題を回避
-assert(fluidFieldSrc.includes('fluidParams') && fluidFieldSrc.includes("from '../config.js'"),
-    'fluid-field.js \u304c fluidParams \u3092 import');
+const toggleKeyMatches = [...devUiSrc.matchAll(/\{\s*key:\s*'([^']+)'/g)].map((m) => m[1]);
+const toggleKeySet = new Set(toggleKeyMatches);
+assert(toggleKeyMatches.length === toggleKeySet.size, 'DEV_TOGGLES の key が重複していない');
 
-assert(distortionSrc.includes('distortionParams') && distortionSrc.includes("from '../config.js'"),
-    'distortion-pass.js \u304c distortionParams \u3092 import');
+section('4. i18n 構造チェック');
 
-assert(/uTintR.*value:\s*sceneParams\.tintR/.test(kessonSrc),
-    'kesson.js: uTintR \u2192 sceneParams.tintR');
-assert(/uTintG.*value:\s*sceneParams\.tintG/.test(kessonSrc),
-    'kesson.js: uTintG \u2192 sceneParams.tintG');
-assert(/uTintB.*value:\s*sceneParams\.tintB/.test(kessonSrc),
-    'kesson.js: uTintB \u2192 sceneParams.tintB');
+const i18nSrc = read(resolve(SRC, 'i18n.js'));
+const topKeys = ['title', 'subtitle', 'toggleLabel', 'taglines', 'credit', 'creditSignature', 'nav', 'gem', 'xLogo'];
 
-// ========================================
-// Test 4: i18n キー構造一致
-// ========================================
-section('4. i18n \u69cb\u9020\u30c1\u30a7\u30c3\u30af');
-
-const i18nSrc = readSrc('i18n.js');
-
-const topKeys = ['title', 'subtitle', 'toggleLabel', 'taglines', 'credit', 'creditSignature', 'nav'];
 for (const key of topKeys) {
     const jaHas = new RegExp(`ja:\\s*\\{[\\s\\S]*?${key}:`).test(i18nSrc);
     const enHas = new RegExp(`en:\\s*\\{[\\s\\S]*?${key}:`).test(i18nSrc);
-    assert(jaHas && enHas, `i18n: ja/en \u4e21\u65b9\u306b '${key}' \u304c\u5b58\u5728`);
+    assert(jaHas && enHas, `i18n: ja/en 両方に '${key}' が存在`);
 }
 
-// ========================================
-// Test 5: 未使用ファイルチェック
-// ========================================
-section('5. \u30d5\u30a1\u30a4\u30eb\u53c2\u7167\u30c1\u30a7\u30c3\u30af');
+section('5. 不要ファイルチェック');
 
-const versionsDir = resolve(SRC, 'versions');
-if (existsSync(versionsDir)) {
-    console.warn('  \u26a0 src/versions/ \u304c\u6b8b\u5b58\uff08\u524a\u9664\u63a8\u5968\uff09');
-} else {
-    passed++;
-}
+assert(!existsSync(resolve(SRC, 'versions')), 'src/versions/ が存在しない');
+assert(!existsSync(resolve(SRC, 'dom-utils.js')), '未使用の src/dom-utils.js が削除済み');
 
-// ========================================
-// 結果サマリー
-// ========================================
-console.log('\n\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550');
-console.log(`\u7d50\u679c: ${passed} passed, ${failed} failed`);
+console.log('\n══════════════════════════════');
+console.log(`結果: ${passed} passed, ${failed} failed`);
 if (failures.length > 0) {
-    console.log('\n\u5931\u6557\u4e00\u89a7:');
+    console.log('\n失敗一覧:');
     failures.forEach((f, i) => console.log(`  ${i + 1}. ${f}`));
 }
-console.log('\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\u2550\n');
+console.log('══════════════════════════════\n');
 
 process.exit(failed > 0 ? 1 : 0);
